Feature/WAIT-6

* [WAIT-6] Adding the new field to the store settings page

* WAIT-6 Remove unwanted line

* WAIT-6 Merge conflicts fixed

* WAIT-6 Type updated


Approved-by: Chris Dawe

+++ grails-app/domain/uk/co/wonderlane/wlpos/PackLine.groovy
@@ -8,7 +8,7 @@ class PackLine {
     String type
     int productListId
     Pack pack
-    int quantity
+    BigDecimal quantity
     String orderCode
 
     static belongsTo = [productListItem: ProductListItem]
